+++ pkg/visor/api.go
@@ -12,6 +12,7 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/bitfield/script"
 	"github.com/ccding/go-stun/stun"
 	"github.com/google/uuid"
 	"github.com/sirupsen/logrus"
@@ -27,6 +28,7 @@ import (
 	"github.com/skycoin/skywire/pkg/transport"
 	"github.com/skycoin/skywire/pkg/transport/network"
 	"github.com/skycoin/skywire/pkg/visor/dmsgtracker"
+	"github.com/skycoin/skywire/pkg/visor/visorconfig"
 )
 
 // API represents visor API.
@@ -36,6 +38,8 @@ type API interface {
 
 	Health() (*HealthInfo, error)
 	Uptime() (float64, error)
+	SetPrivacy(skyenv.Privacy) (string, error)
+	GetPrivacy() (string, error)
 	App(appName string) (*appserver.AppState, error)
 	Apps() ([]*appserver.AppState, error)
 	StartApp(appName string) error
@@ -84,6 +88,8 @@ type API interface {
 
 	GetPersistentTransports() ([]transport.PersistentTransports, error)
 	SetPersistentTransports([]transport.PersistentTransports) error
+	GetLogRotationInterval() (visorconfig.Duration, error)
+	SetLogRotationInterval(visorconfig.Duration) error
 }
 
 // HealthCheckable resource returns its health status as an integer
@@ -301,6 +307,35 @@ func (v *Visor) Health() (*HealthInfo, error) {
 // Uptime implements API.
 func (v *Visor) Uptime() (float64, error) {
 	return time.Since(v.startedAt).Seconds(), nil
+}
+
+// SetPrivacy implements API.
+func (v *Visor) SetPrivacy(p skyenv.Privacy) (string, error) {
+	/*
+		skywire-cli config priv set <address> [flags]
+		Flags:
+		-a, --address string   reward address (default "2jBbGxZRGoQG1mqhPBnXnLTxK6oxsTf8os6")
+		-o, --out string       output config: /opt/skywire/local/privacy.json
+		-i, --publicip         display node ip
+	*/
+	clicmd := `skywire-cli config priv set `
+	//Set flags for node privacy and reward address based on input
+	if p.DisplayNodeIP {
+		clicmd = clicmd + ` -i `
+	}
+	if p.RewardAddress != "" {
+		clicmd = clicmd + ` -a ` + p.RewardAddress
+	}
+	//use the currently configured local_path this visor is using
+	clicmd = clicmd + ` -o ` + strings.Join([]string{v.conf.LocalPath, skyenv.PrivFile}, "/")
+
+	return script.Exec(clicmd).String()
+}
+
+// GetPrivacy implements API.
+func (v *Visor) GetPrivacy() (p string, err error) {
+	clicmd := `skywire-cli config priv get -o ` + strings.Join([]string{v.conf.LocalPath, skyenv.PrivFile}, "/") + ` --json`
+	return script.Exec(clicmd).String()
 }
 
 // Apps implements API.
@@ -627,10 +662,10 @@ func (v *Visor) GetAppConnectionsSummary(appName string) ([]appserver.Connection
 // VPNServers gets available public VPN server from service discovery URL
 func (v *Visor) VPNServers(version, country string) ([]servicedisc.Service, error) {
 	log := logging.MustGetLogger("vpnservers")
-	vlog := logging.NewMasterLogger()
-	vlog.SetLevel(logrus.InfoLevel)
+	vLog := logging.NewMasterLogger()
+	vLog.SetLevel(logrus.InfoLevel)
 
-	sdClient := servicedisc.NewClient(log, vlog, servicedisc.Config{
+	sdClient := servicedisc.NewClient(log, vLog, servicedisc.Config{
 		Type:     servicedisc.ServiceTypeVPN,
 		PK:       v.conf.PK,
 		SK:       v.conf.SK,
@@ -853,9 +888,19 @@ func (v *Visor) SetPersistentTransports(pTps []transport.PersistentTransports) e
 	return v.conf.UpdatePersistentTransports(pTps)
 }
 
-// GetPersistentTransports sets min_hops routing config of visor
+// GetPersistentTransports gets min_hops routing config of visor
 func (v *Visor) GetPersistentTransports() ([]transport.PersistentTransports, error) {
 	return v.conf.GetPersistentTransports()
+}
+
+// SetLogRotationInterval sets log_rotation_interval config of visor
+func (v *Visor) SetLogRotationInterval(d visorconfig.Duration) error {
+	return v.conf.UpdateLogRotationInterval(d)
+}
+
+// GetLogRotationInterval gets log_rotation_interval config of visor
+func (v *Visor) GetLogRotationInterval() (visorconfig.Duration, error) {
+	return v.conf.GetLogRotationInterval()
 }
 
 // SetPublicAutoconnect sets public_autoconnect config of visor